feat(aws-secret-provider): Make separate package for aws
