+++ bundle/camunda-saas-bundle/pom.xml
@@ -30,15 +30,20 @@
       </exclusions>
     </dependency>
     <dependency>
+      <groupId>io.camunda.connector</groupId>
+      <artifactId>aws-secret-provider</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>io.camunda.connector</groupId>
+      <artifactId>gcp-secret-provider</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-oauth2-resource-server</artifactId>
     </dependency>
 
-    <dependency>
-      <groupId>io.camunda.connector</groupId>
-      <artifactId>connector-gcp-secret-provider</artifactId>
-      <version>${project.version}</version>
-    </dependency>
     <dependency>
       <groupId>com.google.cloud</groupId>
       <artifactId>spring-cloud-gcp-starter-logging</artifactId>
@@ -69,12 +74,6 @@
       <groupId>org.springframework.security</groupId>
       <artifactId>spring-security-test</artifactId>
       <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.camunda.connector</groupId>
-      <artifactId>secret-provider</artifactId>
-      <version>8.8.0-SNAPSHOT</version>
-      <scope>compile</scope>
     </dependency>
   </dependencies>
