+++ secret-providers/aws/src/main/resources/io.camunda.connector.api.secret.SecretProvider
@@ -1,0 +1,1 @@
+io.camunda.connector.secret.providers.AwsSecretProvider
