--- secret-providers/src/main/resources/META-INF/services/io.camunda.connector.api.secret.SecretProvider
@@ -1,2 +1,0 @@
-io.camunda.connector.secret.provider.GcpSecretProvider
-io.camunda.connector.secret.provider.AwsSecretProvider
